Extract cognition, opinion, and reported speech
