+++ main.js
@@ -8363,6 +8363,658 @@ var require_question_fallbacks = __commonJS({
   }
 });
 
+// src/parser/detectors/cognition/fragments-content.js
+var require_fragments_content = __commonJS({
+  "src/parser/detectors/cognition/fragments-content.js"(exports2, module2) {
+    "use strict";
+    module2.exports = function createCognitionDetectors2(dependencies = {}) {
+      const {
+        aNotAQuestionFallback: aNotAQuestionFallback2,
+        applyConstructionPatterns: applyConstructionPatterns2,
+        completionQuestionFallback: completionQuestionFallback2,
+        construction: construction2,
+        desiderativeANotAQuestionFallback: desiderativeANotAQuestionFallback2,
+        flattenSurface: flattenSurface2,
+        isToken: isToken2,
+        nodeCanFillSlot: nodeCanFillSlot2,
+        nodeSurfaceMatches: nodeSurfaceMatches2,
+        optionalSubjectOffset: optionalSubjectOffset2,
+        permissionANotAQuestionFallback: permissionANotAQuestionFallback2,
+        phase4CognitionActiveTokenClone: phase4CognitionActiveTokenClone2,
+        predicateOmissionProfileForHead: predicateOmissionProfileForHead2,
+        templateConstructionFor: templateConstructionFor2,
+        templateDerivedSlots: templateDerivedSlots2,
+        token: token2,
+        traceInfo: traceInfo2,
+        withoutTrailingParticles: withoutTrailingParticles2,
+        wrapCategorySubspans: wrapCategorySubspans2
+      } = dependencies;
+      function negativeCognitionFragmentFallback2(core) {
+        const { core: bareCore, particles } = withoutTrailingParticles2(core);
+        let negator = null;
+        let predicate = null;
+        if (bareCore.length === 2 && nodeCanFillSlot2(bareCore[0], "negator") && nodeCanFillSlot2(bareCore[1], "cognition_predicate")) {
+          negator = bareCore[0];
+          predicate = bareCore[1];
+        } else if (bareCore.length === 1 && flattenSurface2(bareCore[0]) === "唔知") {
+          negator = token2("唔");
+          predicate = token2("知");
+        } else {
+          return null;
+        }
+        if (!nodeCanFillSlot2(negator, "negator")) return null;
+        if (!nodeCanFillSlot2(predicate, "cognition_predicate")) return null;
+        const promoted = phase4CognitionActiveTokenClone2(predicate, {
+          label: predicate.label || "doing",
+          syntax: `${predicate.syntax || "cognition_verb"} negative_cognition_predicate`,
+          slots: ["negative_cognition_predicate", "negative_cognition_fragment"],
+          reason: "Phase 4 controlled grammar promotion: 知 is parser-active only inside the reviewed 唔知 negative-cognition fragment; 唔 remains a learner-visible func negator."
+        });
+        const children = [negator, promoted, ...particles];
+        return construction2("NegativeCognitionFragment", "NegCognition", children, {
+          note: "Phase 4 controlled negative cognition fragment: 唔 + 知 as a stable short answer / learner fragment, with 唔 kept as func.",
+          slots: templateDerivedSlots2("NegativeCognitionFragment", children),
+          trace: traceInfo2("phase4_controlled_grammar_promotion", {
+            rule: "negator + cognition_predicate + particle?",
+            reason: "Promote only reviewed 唔知-style negative cognition fragments. The negator remains a learner-visible func token; this does not promote general cognition, speech, or stance domains.",
+            surfaces: children.map((node) => flattenSurface2(node))
+          })
+        });
+      }
+      function cognitionStatementFallback2(core) {
+        const { core: bareCore, particles } = withoutTrailingParticles2(core);
+        if (bareCore.length !== 2) return null;
+        const subject = bareCore[0];
+        const predicate = bareCore[1];
+        if (!nodeCanFillSlot2(subject, "subject")) return null;
+        if (!nodeCanFillSlot2(predicate, "cognition_predicate")) return null;
+        if (nodeCanFillSlot2(predicate, "negative_cognition_predicate")) return null;
+        if (nodeCanFillSlot2(predicate, "stance_predicate")) return null;
+        const promoted = phase4CognitionActiveTokenClone2(predicate, {
+          label: predicate.label || "doing",
+          syntax: `${predicate.syntax || "cognition_verb"} cognition_statement_predicate`,
+          slots: ["cognition_statement_clause"],
+          reason: "Phase 4 controlled grammar promotion: bare subject + 知 is parser-active as a reviewed cognition statement, not a broad cognition-content frame."
+        });
+        const children = [subject, promoted, ...particles];
+        return construction2("CognitionStatementClause", "Know", children, {
+          note: "Phase 4 controlled cognition statement: subject + 知 + optional particle.",
+          slots: templateDerivedSlots2("CognitionStatementClause", children),
+          trace: traceInfo2("phase4_controlled_grammar_promotion", {
+            rule: "subject + non_stance_cognition_predicate + particle?",
+            reason: "Promote only reviewed bare cognition statements such as 我知 / 我知呀. Content frames still require visible clause-like content, and desiderative 我想知 remains separate.",
+            predicate_omission_profile: "factive_cognition",
+            omission_status: "conventional_bare_statement",
+            complement_type: "proposition_or_interrogative_clause",
+            conventionality_status: "conventional_bare_statement",
+            polarity: "positive",
+            speech_event_use: "not_applicable",
+            not_claims: ["not_fabricated_content_clause", "not_forced_fragment_answer"],
+            surfaces: children.map((node) => flattenSurface2(node))
+          })
+        });
+      }
+      function cognitionContentFrameFallback2(core) {
+        let cursor = optionalSubjectOffset2(core);
+        const subjectNodes = core.slice(0, cursor);
+        const negator = isToken2(core[cursor], "唔") ? core[cursor++] : null;
+        if (core.length - cursor < 2) return null;
+        const cognition = core[cursor];
+        if (!nodeCanFillSlot2(cognition, "cognition_predicate")) return null;
+        if (nodeCanFillSlot2(cognition, "negative_cognition_predicate")) return null;
+        if (nodeCanFillSlot2(cognition, "stance_predicate")) return null;
+        if (nodeCanFillSlot2(cognition, "vp_complement_predicate") && !nodeSurfaceMatches2(cognition, ["知"])) return null;
+        const rawContent = core.slice(cursor + 1);
+        const { core: contentCore, particles } = withoutTrailingParticles2(rawContent);
+        const directClauseRelationEdge = contentCore.length === 1 && contentCore[0].kind === "construction" && contentCore[0].type === "ClauseRelationEdge" ? contentCore[0] : null;
+        if (!directClauseRelationEdge && contentCore.length < 2) return null;
+        const rawEmbeddedQuestion = directClauseRelationEdge ? null : aNotAQuestionFallback2(contentCore) || permissionANotAQuestionFallback2(contentCore) || desiderativeANotAQuestionFallback2(contentCore) || completionQuestionFallback2(contentCore);
+        const parsedContent = directClauseRelationEdge ? [directClauseRelationEdge] : rawEmbeddedQuestion ? [rawEmbeddedQuestion] : applyConstructionPatterns2(contentCore);
+        const wrappedContent = parsedContent.length === 1 ? parsedContent : wrapCategorySubspans2(contentCore);
+        let contentNode = directClauseRelationEdge;
+        if (!contentNode && (wrappedContent.length === 1 && wrappedContent[0].kind === "construction" && (nodeCanFillSlot2(wrappedContent[0], "reported_content") || nodeCanFillSlot2(wrappedContent[0], "clause") || nodeCanFillSlot2(wrappedContent[0], "vp") || nodeCanFillSlot2(wrappedContent[0], "predicate") || nodeCanFillSlot2(wrappedContent[0], "question_fragment")))) {
+          contentNode = wrappedContent[0];
+        } else if (!contentNode) {
+          const generatedContent = templateConstructionFor2(wrappedContent, [
+            "ANotAQuestion",
+            "OpinionStanceFrame",
+            "ReportedSpeech",
+            "ExistentialClause",
+            "NegatedExistentialClause",
+            "ExistentialQuestion",
+            "DesiderativeVP",
+            "ModalVP",
+            "CompletionQuestion"
+          ]);
+          if (generatedContent && (nodeCanFillSlot2(generatedContent, "reported_content") || nodeCanFillSlot2(generatedContent, "clause") || nodeCanFillSlot2(generatedContent, "vp") || nodeCanFillSlot2(generatedContent, "predicate") || nodeCanFillSlot2(generatedContent, "question_fragment"))) contentNode = generatedContent;
+        }
+        if (!contentNode) return null;
+        const promotedCognition = phase4CognitionActiveTokenClone2(cognition, {
+          label: cognition.label || "doing",
+          syntax: `${cognition.syntax || "cognition_verb"} non_stance_cognition_predicate`,
+          slots: ["non_stance_cognition_predicate"],
+          reason: "A1 broad cognition architecture: a 知-type predicate is parser-active when it licenses a visible declarative or interrogative content complement."
+        });
+        const children = [...subjectNodes, ...negator ? [negator] : [], promotedCognition, contentNode, ...particles];
+        const profile = predicateOmissionProfileForHead2(flattenSurface2(cognition));
+        return construction2("CognitionContentFrame", "Know+Content", children, {
+          note: "Cognition content frame with visible polarity and a transparent declarative or interrogative complement.",
+          slots: templateDerivedSlots2("CognitionContentFrame", children),
+          trace: traceInfo2("phase4_controlled_grammar_promotion", {
+            rule: "subject? + polarity? + non_stance_cognition_predicate + content_clause + particle?",
+            reason: "The content remains an overt child construction. Negative polarity composes outside the cognition head, and embedded A-not-A content is not flattened into multiple roots.",
+            content_construction: contentNode.type,
+            predicate_omission_profile: profile ? profile.id : "factive_cognition",
+            omission_status: "overt_content_saturated",
+            complement_type: contentNode.type && String(contentNode.type).includes("Question") ? "interrogative_clause" : "proposition",
+            conventionality_status: "overt_content",
+            polarity: negator ? "negative" : "positive",
+            speech_event_use: "not_applicable",
+            surfaces: children.map((node) => flattenSurface2(node)),
+            not_claims: ["not_hidden_content_clause", "not_stance_frame", "not_reported_speech"]
+          })
+        });
+      }
+      return {
+        cognitionContentFrameFallback: cognitionContentFrameFallback2,
+        cognitionStatementFallback: cognitionStatementFallback2,
+        negativeCognitionFragmentFallback: negativeCognitionFragmentFallback2
+      };
+    };
+  }
+});
+
+// src/parser/detectors/opinion/stance.js
+var require_stance = __commonJS({
+  "src/parser/detectors/opinion/stance.js"(exports2, module2) {
+    "use strict";
+    module2.exports = function createOpinionDetectors2(dependencies = {}) {
+      const {
+        applyConstructionPatterns: applyConstructionPatterns2,
+        categorySubspanFor: categorySubspanFor2,
+        cleanSlots: cleanSlots2,
+        construction: construction2,
+        copulaClone: copulaClone2,
+        firstToken: firstToken2,
+        flattenSurface: flattenSurface2,
+        isToken: isToken2,
+        modalVPFromNodes: modalVPFromNodes2,
+        nodeCanFillSlot: nodeCanFillSlot2,
+        nodeSlots: nodeSlots2,
+        nominalComplementFromNodes: nominalComplementFromNodes2,
+        parserInactiveTokenClone: parserInactiveTokenClone2,
+        phase4OpinionStanceActiveTokenClone: phase4OpinionStanceActiveTokenClone2,
+        predicateOmissionProfileForHead: predicateOmissionProfileForHead2,
+        subjectStativePredicateClauseFallback: subjectStativePredicateClauseFallback2,
+        templateConstructionFor: templateConstructionFor2,
+        templateDerivedSlots: templateDerivedSlots2,
+        traceInfo: traceInfo2,
+        withoutTrailingParticles: withoutTrailingParticles2,
+        wrapCategorySubspans: wrapCategorySubspans2
+      } = dependencies;
+      function splitOpinionTrailingParticles(core) {
+        const split = withoutTrailingParticles2(core);
+        const bareCore = split.core.slice();
+        const particles = split.particles.slice();
+        if (particles.length && isToken2(particles[0], "嘅") && bareCore.some((node) => isToken2(node, "係"))) {
+          bareCore.push(particles.shift());
+        }
+        return { core: bareCore, particles };
+      }
+      function isOpinionFrameFocusModifier(node) {
+        return Boolean(node) && node.kind !== "text" && nodeCanFillSlot2(node, "focus_adverb");
+      }
+      function splitLeadingOpinionContentModifiers(nodes) {
+        const modifiers = [];
+        let index = 0;
+        while (index < nodes.length && isOpinionFrameFocusModifier(nodes[index])) {
+          modifiers.push(nodes[index]);
+          index += 1;
+        }
+        return { modifiers, core: nodes.slice(index) };
+      }
+      function opinionComplementTokenClone(node, subtype) {
+        if (!node || node.kind !== "token") return node;
+        const surface = flattenSurface2(node);
+        if (["係", "唔係"].includes(surface)) {
+          return parserInactiveTokenClone2(node, {
+            label: "func",
+            pos: "function",
+            syntax: `${surface === "唔係" ? "negated_copula" : "copula"} complement_ellipsis_predicate`,
+            slots: cleanSlots2([
+              surface === "唔係" ? "negated_copula" : "copula",
+              surface === "唔係" ? "negator" : "",
+              "predicate",
+              "reported_content",
+              "content_clause"
+            ]),
+            reason: `The visible ${surface} predicate is retained inside a ${subtype} complement-ellipsis fragment; no missing complement words are fabricated.`
+          });
+        }
+        if (["有", "冇"].includes(surface)) {
+          return parserInactiveTokenClone2(node, {
+            label: "func",
+            pos: "function",
+            syntax: `${surface === "冇" ? "negated_existential" : "existential"} complement_ellipsis_predicate`,
+            slots: cleanSlots2([
+              surface === "冇" ? "negated_existential" : "existential",
+              "predicate",
+              "reported_content",
+              "content_clause"
+            ]),
+            reason: `The visible ${surface} predicate is retained inside an existential-domain complement-ellipsis fragment; no missing domain is fabricated.`
+          });
+        }
+        return node;
+      }
+      function opinionComplementEllipsisFragment(nodes, subtype, missingSlot) {
+        const children = nodes.map((node) => opinionComplementTokenClone(node, subtype));
+        return construction2("ComplementEllipsisFragment", "ComplementEllipsis", children, {
+          slots: cleanSlots2([
+            "complement_ellipsis_fragment",
+            "fragment_answer",
+            "reported_content",
+            "content_clause",
+            "predicate",
+            "clause",
+            ...children.flatMap((node) => nodeSlots2(node))
+          ]),
+          note: "Visible opinion content with a typed complement/domain omitted and recoverable only from discourse.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "ComplementEllipsisFragment",
+            template_family: "generative_template",
+            template: subtype === "existential_domain_ellipsis" ? ["existential_predicate!"] : nodes.length > 1 ? ["subject?", "copular_predicate!"] : ["copular_predicate!"],
+            assigned_slots: nodes.map((node, index) => {
+              if (index === 0 && nodes.length > 1 && nodeCanFillSlot2(node, "subject")) return "subject";
+              return subtype === "existential_domain_ellipsis" ? "existential_predicate" : "copular_predicate";
+            }),
+            fragment_subtype: subtype,
+            context_requirement_status: "context_required",
+            missing_argument_slots: [missingSlot],
+            missing_slot_details: [{ slot: missingSlot, license_status: "unresolved" }],
+            antecedent_status: "not_observed",
+            discourse_license_not_observed: true,
+            surfaces: children.map((node) => flattenSurface2(node)),
+            reason: "The overt predicate is structurally valid, but its selected complement/domain is absent and must remain linked to discourse rather than being invented.",
+            not_claims: ["not_context_free_complete_clause", "not_fabricated_complement", "not_lexicalized_whole_sentence"]
+          })
+        });
+      }
+      function nominalizedStativeComplementForOpinion(nodes) {
+        if (!nodes || nodes.length !== 2) return null;
+        const stative = nodes[0];
+        const linker = nodes[1];
+        if (!nodeCanFillSlot2(stative, "stative_predicate") || !isToken2(linker, "嘅")) return null;
+        const nominalizer = parserInactiveTokenClone2(linker, {
+          label: "func",
+          pos: "function",
+          syntax: "nominalizer nominal_linker",
+          slots: ["nominal_linker"],
+          reason: "嘅 nominalizes the visible stative predicate inside a copular opinion complement; it is not treated as a sentence-final particle."
+        });
+        const children = [stative, nominalizer];
+        return construction2("StativeNominalComplement", "AdjNP", children, {
+          slots: cleanSlots2([
+            "stative_nominal_complement",
+            "copular_complement",
+            "np",
+            "predicate",
+            "stative_predicate",
+            "nominal_linker",
+            ...children.flatMap((node) => nodeSlots2(node))
+          ]),
+          note: "Headless nominalized stative complement such as 真嘅 in 係真嘅.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "StativeNominalComplement",
+            template_family: "generative_template",
+            template: ["stative_predicate!", "nominal_linker!"],
+            assigned_slots: ["stative_predicate", "nominal_linker"],
+            complement_subtype: "headless_nominalized_stative",
+            surfaces: children.map((node) => flattenSurface2(node)),
+            reason: "A visible stative predicate plus nominalizer forms the copular complement without inventing a silent noun head.",
+            not_claims: ["not_overt_head_noun", "not_sentence_final_particle_use"]
+          })
+        });
+      }
+      function opinionCopularContentNode(nodes) {
+        if (!nodes || nodes.length < 2 || !isToken2(nodes[0], "係")) return null;
+        const complementNodes = nodes.slice(1);
+        const complement = nominalizedStativeComplementForOpinion(complementNodes) || nominalComplementFromNodes2(complementNodes) || categorySubspanFor2(complementNodes, ["DegreeStativePredicate", "NegatedStativePredicate", "StativePredicate"]);
+        if (!complement) return null;
+        const copula = copulaClone2(
+          nodes[0],
+          "copula subordinate_opinion_copula",
+          ["copula"],
+          "係 heads a visible copular proposition inside the opinion complement. Its understood proposition subject is licensed by the embedding stance frame, not fabricated as a token."
+        );
+        const children = [copula, complement];
+        return construction2("CopularRelationFrame", "Relation", children, {
+          slots: cleanSlots2([
+            "copular_relation_frame",
+            "copular_clause",
+            "copula",
+            "copular_complement",
+            "reported_content",
+            "content_clause",
+            "predicate",
+            "clause",
+            ...children.flatMap((node) => nodeSlots2(node))
+          ]),
+          note: "Subordinate copular proposition used as visible opinion content.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "CopularRelationFrame",
+            template_family: "generative_template",
+            template: ["copula!", "copular_complement!"],
+            assigned_slots: ["copula", "copular_complement"],
+            subject_status: "null_licensed_by_opinion_content",
+            context_requirement_status: "context_not_required",
+            missing_argument_slots: [],
+            antecedent_status: "not_applicable",
+            surfaces: children.map((node) => flattenSurface2(node)),
+            reason: "Inside a stance complement, 係 plus a visible nominalized/stative complement forms a complete proposition without requiring an invented overt subject.",
+            not_claims: ["not_standalone_context_free_subject_claim", "not_complement_ellipsis"]
+          })
+        });
+      }
+      function contentNodeForOpinionStanceFrame(contentCore) {
+        if (!contentCore.length) return null;
+        const split = splitLeadingOpinionContentModifiers(contentCore);
+        const leadingModifiers = split.modifiers;
+        const core = split.core;
+        if (!core.length) return null;
+        if (core.length === 1 && core[0].kind === "construction" && core[0].type === "ClauseRelationEdge") {
+          return { leadingModifiers, contentNode: core[0] };
+        }
+        if (core.length === 1 && ["有", "冇"].includes(flattenSurface2(core[0]))) {
+          return {
+            leadingModifiers,
+            contentNode: opinionComplementEllipsisFragment(core, "existential_domain_ellipsis", "existential_domain")
+          };
+        }
+        const finalSurface = flattenSurface2(core[core.length - 1]);
+        const copularEllipsis = ["係", "唔係"].includes(finalSurface) && (core.length === 1 || core.length === 2 && nodeCanFillSlot2(core[0], "subject"));
+        if (copularEllipsis) {
+          return {
+            leadingModifiers,
+            contentNode: opinionComplementEllipsisFragment(core, "copular_complement_ellipsis", "copular_complement")
+          };
+        }
+        const copularContent = opinionCopularContentNode(core);
+        if (copularContent) return { leadingModifiers, contentNode: copularContent };
+        const modalContent = modalVPFromNodes2(core);
+        if (modalContent) return { leadingModifiers, contentNode: modalContent };
+        const stativeContent = categorySubspanFor2(core, [
+          "DegreeStativePredicate",
+          "NegatedStativePredicate",
+          "DegreeModifiedLexicalStative",
+          "StativePredicate"
+        ]);
+        if (stativeContent) return { leadingModifiers, contentNode: stativeContent };
+        const fullyWrappedContent = applyConstructionPatterns2(core);
+        if (fullyWrappedContent.length === 1 && fullyWrappedContent[0].kind === "construction" && [
+          "ClauseSpan",
+          "TopicComment",
+          "CopularRelationFrame",
+          "CopularIdentificationFrame",
+          "ExistentialClause",
+          "NegatedExistentialClause",
+          "DesiderativeVP",
+          "ModalVP",
+          "CompletionQuestion",
+          "ANotAQuestion",
+          "PotentialResultVP",
+          "PerfectiveVP",
+          "PostverbalZoPerfectiveVP",
+          "DirectionalMotionVP",
+          "DegreeStativePredicate",
+          "NegatedStativePredicate",
+          "StativePredicate",
+          "ComplementEllipsisFragment",
+          "ClauseRelationEdge"
+        ].includes(fullyWrappedContent[0].type)) {
+          return { leadingModifiers, contentNode: fullyWrappedContent[0] };
+        }
+        const wrappedContent = wrapCategorySubspans2(core);
+        const subjectStative = subjectStativePredicateClauseFallback2(wrappedContent);
+        if (subjectStative) return { leadingModifiers, contentNode: subjectStative };
+        if (wrappedContent.length === 1 && wrappedContent[0].kind === "construction" && [
+          "ClauseSpan",
+          "TopicComment",
+          "CopularRelationFrame",
+          "CopularIdentificationFrame",
+          "ExistentialClause",
+          "NegatedExistentialClause",
+          "DesiderativeVP",
+          "ModalVP",
+          "CompletionQuestion",
+          "DegreeStativePredicate",
+          "NegatedStativePredicate",
+          "StativePredicate",
+          "ComplementEllipsisFragment",
+          "ClauseRelationEdge"
+        ].includes(wrappedContent[0].type)) {
+          return { leadingModifiers, contentNode: wrappedContent[0] };
+        }
+        const generatedContent = templateConstructionFor2(wrappedContent, [
+          "ExistentialClause",
+          "NegatedExistentialClause",
+          "DesiderativeVP",
+          "ModalVP",
+          "CompletionQuestion"
+        ]);
+        if (generatedContent) return { leadingModifiers, contentNode: generatedContent };
+        return null;
+      }
+      function opinionStanceFrameFallback2(core) {
+        const split = splitOpinionTrailingParticles(core);
+        const bareCore = split.core;
+        const particles = split.particles;
+        if (bareCore.length < 3) return null;
+        let cursor = 0;
+        const subject = nodeCanFillSlot2(bareCore[cursor], "subject") ? bareCore[cursor++] : null;
+        const preStanceModifiers = [];
+        while (cursor < bareCore.length && isOpinionFrameFocusModifier(bareCore[cursor])) {
+          preStanceModifiers.push(bareCore[cursor]);
+          cursor += 1;
+        }
+        const stancePredicate = bareCore[cursor];
+        const stanceToken = firstToken2(stancePredicate);
+        if (!stanceToken || !nodeCanFillSlot2(stancePredicate, "stance_predicate")) return null;
+        const profile = predicateOmissionProfileForHead2(stanceToken.surface);
+        cursor += 1;
+        if (cursor >= bareCore.length) return null;
+        const rawStanceContent = bareCore.slice(cursor);
+        const contentAnalysis = rawStanceContent.length === 1 && rawStanceContent[0].kind === "construction" && rawStanceContent[0].type === "ClauseRelationEdge" ? { leadingModifiers: [], contentNode: rawStanceContent[0] } : contentNodeForOpinionStanceFrame(rawStanceContent);
+        if (!contentAnalysis || !contentAnalysis.contentNode) return null;
+        const contentNode = contentAnalysis.contentNode;
+        const postStanceModifiers = contentAnalysis.leadingModifiers || [];
+        const promotedStance = phase4OpinionStanceActiveTokenClone2(stancePredicate, {
+          label: stancePredicate.label || "doing",
+          syntax: `${stanceToken.syntax || "cognition_opinion_predicate"} opinion_stance_predicate`,
+          slots: ["opinion_stance_frame"],
+          reason: "A1 broad stance architecture: a registered stance or belief predicate is parser-active when it licenses visible reviewed content."
+        });
+        const children = [
+          ...subject ? [subject] : [],
+          ...preStanceModifiers,
+          promotedStance,
+          ...postStanceModifiers,
+          contentNode,
+          ...particles
+        ];
+        const assignedSlots = [
+          ...subject ? ["subject"] : [],
+          ...preStanceModifiers.map(() => "focus_adverb"),
+          "stance_predicate",
+          ...postStanceModifiers.map(() => "focus_adverb"),
+          "reported_content",
+          ...particles.map(() => "particle")
+        ];
+        const contentTrace = contentNode.trace || {};
+        const contextRequirementStatus = contentTrace.context_requirement_status || "context_not_required";
+        const missingArgumentSlots = Array.isArray(contentTrace.missing_argument_slots) ? contentTrace.missing_argument_slots.slice() : [];
+        const missingSlotDetails = Array.isArray(contentTrace.missing_slot_details) ? contentTrace.missing_slot_details.map((item) => ({ ...item })) : missingArgumentSlots.map((slot) => ({ slot, license_status: "unresolved" }));
+        return construction2("OpinionStanceFrame", "Opinion/Stance", children, {
+          note: "Opinion, evaluation, or belief frame with a visible transparent proposition or typed complement ellipsis.",
+          slots: cleanSlots2([
+            ...templateDerivedSlots2("OpinionStanceFrame", children),
+            "opinion_stance_frame",
+            "stance_predicate",
+            "reported_content",
+            "content_clause",
+            "predicate",
+            "clause"
+          ]),
+          trace: traceInfo2("generative_template", {
+            construction_type: "OpinionStanceFrame",
+            template_family: "generative_template",
+            template: ["subject?", "focus_adverb?", "stance_predicate!", "focus_adverb?", "reported_content!", "particle?"],
+            constraints: {
+              visible_reviewed_content: true,
+              stance_predicate_surfaces: [stanceToken.surface],
+              focus_modifier_slot: "focus_adverb"
+            },
+            assigned_slots: assignedSlots,
+            content_construction: contentNode.type,
+            pre_stance_modifier_surfaces: preStanceModifiers.map((node) => flattenSurface2(node)),
+            post_stance_modifier_surfaces: postStanceModifiers.map((node) => flattenSurface2(node)),
+            context_requirement_status: contextRequirementStatus,
+            missing_argument_slots: missingArgumentSlots,
+            missing_slot_details: missingSlotDetails,
+            antecedent_status: contextRequirementStatus === "context_required" ? "not_observed" : "not_applicable",
+            discourse_license_not_observed: contextRequirementStatus === "context_required",
+            attached_particles: particles.map((node) => flattenSurface2(node)),
+            surfaces: children.map((node) => flattenSurface2(node)),
+            predicate_omission_profile: profile ? profile.id : "opinion_evaluation",
+            reason: "The reusable stance frame is keyed by registered stance-predicate affordance, not one hard-coded surface. It wraps—but does not replace—the visible content construction.",
+            not_claims: ["not_reported_speech", "not_action_stative_vp_overcapture", "not_fabricated_opinion_content"]
+          })
+        });
+      }
+      return { opinionStanceFrameFallback: opinionStanceFrameFallback2 };
+    };
+  }
+});
+
+// src/parser/detectors/reported-speech/composition.js
+var require_composition = __commonJS({
+  "src/parser/detectors/reported-speech/composition.js"(exports2, module2) {
+    "use strict";
+    module2.exports = function createReportedSpeechDetectors2(dependencies = {}) {
+      const {
+        applyConstructionPatterns: applyConstructionPatterns2,
+        construction: construction2,
+        firstToken: firstToken2,
+        flattenSurface: flattenSurface2,
+        nodeCanFillSlot: nodeCanFillSlot2,
+        optionalSubjectOffset: optionalSubjectOffset2,
+        parserInactiveTokenClone: parserInactiveTokenClone2,
+        phase4ReportedSpeechActiveTokenClone: phase4ReportedSpeechActiveTokenClone2,
+        subjectStativePredicateClauseFallback: subjectStativePredicateClauseFallback2,
+        templateConstructionFor: templateConstructionFor2,
+        templateDerivedSlots: templateDerivedSlots2,
+        traceInfo: traceInfo2,
+        withoutTrailingParticles: withoutTrailingParticles2,
+        wrapCategorySubspans: wrapCategorySubspans2
+      } = dependencies;
+      function contentNodeForReportedSpeech(contentCore) {
+        if (!contentCore.length) return null;
+        if (contentCore.length === 1 && contentCore[0].kind === "construction" && contentCore[0].type === "ClauseRelationEdge") {
+          return contentCore[0];
+        }
+        if (contentCore.length === 1 && contentCore[0].kind === "token") {
+          const contentToken = contentCore[0];
+          if (["有", "冇"].includes(flattenSurface2(contentToken)) && (nodeCanFillSlot2(contentToken, "existential") || nodeCanFillSlot2(contentToken, "negated_existential"))) {
+            return parserInactiveTokenClone2(contentToken, {
+              label: contentToken.label || contentToken.role || "func",
+              syntax: `${contentToken.syntax || "predicate"} reported_content_predicate`,
+              slots: [...contentToken.slots || [], "reported_content", "predicate"],
+              reason: "A visible short predicate is functioning as the reported content of 話; no standalone response-fragment reading or hidden clause is introduced."
+            });
+          }
+        }
+        const parsedContent = applyConstructionPatterns2(contentCore);
+        const wrappedContent = parsedContent.length === 1 ? parsedContent : wrapCategorySubspans2(contentCore);
+        const subjectStative = subjectStativePredicateClauseFallback2(wrappedContent);
+        if (subjectStative) return subjectStative;
+        if (wrappedContent.length === 1 && wrappedContent[0].kind === "construction" && (nodeCanFillSlot2(wrappedContent[0], "reported_content") || nodeCanFillSlot2(wrappedContent[0], "clause") || nodeCanFillSlot2(wrappedContent[0], "vp") || nodeCanFillSlot2(wrappedContent[0], "predicate"))) {
+          return wrappedContent[0];
+        }
+        if (wrappedContent.length === 1 && wrappedContent[0].kind === "token") {
+          const contentToken = wrappedContent[0];
+          if (nodeCanFillSlot2(contentToken, "predicate") || nodeCanFillSlot2(contentToken, "existential") || nodeCanFillSlot2(contentToken, "negated_existential") || nodeCanFillSlot2(contentToken, "vp") || nodeCanFillSlot2(contentToken, "clause")) {
+            return parserInactiveTokenClone2(contentToken, {
+              label: contentToken.label || contentToken.role || "func",
+              syntax: `${contentToken.syntax || "predicate"} reported_content_predicate`,
+              slots: [...contentToken.slots || [], "reported_content", "predicate"],
+              reason: "A visible short predicate is functioning as the reported content of 話; no hidden clause or lexicalized whole-sentence formula is introduced."
+            });
+          }
+        }
+        const generatedContent = templateConstructionFor2(wrappedContent, [
+          "CognitionContentFrame",
+          "OpinionStanceFrame",
+          "ExistentialClause",
+          "NegatedExistentialClause",
+          "DesiderativeVP",
+          "ModalVP",
+          "CompletionQuestion"
+        ]);
+        if (generatedContent && (nodeCanFillSlot2(generatedContent, "reported_content") || nodeCanFillSlot2(generatedContent, "clause") || nodeCanFillSlot2(generatedContent, "vp") || nodeCanFillSlot2(generatedContent, "predicate"))) return generatedContent;
+        return null;
+      }
+      function reportedSpeechFrameFallback2(core) {
+        const { core: bareCore, particles } = withoutTrailingParticles2(core);
+        const offset = optionalSubjectOffset2(bareCore);
+        if (bareCore.length - offset < 2) return null;
+        const speechVerb = bareCore[offset];
+        const speechToken = firstToken2(speechVerb);
+        if (!speechToken || speechToken.surface !== "話") return null;
+        if (!nodeCanFillSlot2(speechVerb, "speech_verb")) return null;
+        const contentNode = contentNodeForReportedSpeech(bareCore.slice(offset + 1));
+        if (!contentNode) return null;
+        const promotedSpeech = phase4ReportedSpeechActiveTokenClone2(speechVerb, {
+          label: speechVerb.label || "doing",
+          syntax: `${speechToken.syntax || "speech_reporting_verb"} reported_speech_predicate`,
+          slots: ["reported_speech"],
+          reason: "Phase 4 controlled grammar promotion: 話 is parser-active only when it introduces a visible reviewed reported-content frame."
+        });
+        const children = [
+          ...bareCore.slice(0, offset),
+          promotedSpeech,
+          contentNode,
+          ...particles
+        ];
+        const assignedSlots = [
+          ...bareCore.slice(0, offset).map(() => "subject"),
+          "speech_verb",
+          "reported_content",
+          ...particles.map(() => "particle")
+        ];
+        return construction2("ReportedSpeech", "Reported", children, {
+          note: "Reported speech frame: optional subject + speech-reporting verb + visible reviewed reported content.",
+          slots: templateDerivedSlots2("ReportedSpeech", children),
+          trace: traceInfo2("generative_template", {
+            construction_type: "ReportedSpeech",
+            template_family: "generative_template",
+            template: ["subject?", "speech_verb!", "reported_content!", "particle?"],
+            constraints: {
+              visible_reviewed_content: true,
+              speech_verb_surfaces: ["話"]
+            },
+            assigned_slots: assignedSlots,
+            content_construction: contentNode.type,
+            attached_particles: particles.map((node) => flattenSurface2(node)),
+            surfaces: children.map((node) => flattenSurface2(node)),
+            reason: "The frame is now reported as a named reusable generative template; the controlled parser-active speech predicate remains scoped to this approved construction."
+          })
+        });
+      }
+      return { reportedSpeechFrameFallback: reportedSpeechFrameFallback2 };
+    };
+  }
+});
+
 // src/parser/detectors/discourse/formula-responses.js
 var require_formula_responses = __commonJS({
   "src/parser/detectors/discourse/formula-responses.js"(exports2, module2) {
@@ -11951,38 +12603,73 @@ var { haveOrNotQuestionFallbackForPunctuation } = createQuestionFallbacks({
   templateDerivedSlots,
   traceInfo
 });
-function negativeCognitionFragmentFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  let negator = null;
-  let predicate = null;
-  if (bareCore.length === 2 && nodeCanFillSlot(bareCore[0], "negator") && nodeCanFillSlot(bareCore[1], "cognition_predicate")) {
-    negator = bareCore[0];
-    predicate = bareCore[1];
-  } else if (bareCore.length === 1 && flattenSurface(bareCore[0]) === "唔知") {
-    negator = token("唔");
-    predicate = token("知");
-  } else {
-    return null;
-  }
-  if (!nodeCanFillSlot(negator, "negator")) return null;
-  if (!nodeCanFillSlot(predicate, "cognition_predicate")) return null;
-  const promoted = phase4CognitionActiveTokenClone(predicate, {
-    label: predicate.label || "doing",
-    syntax: `${predicate.syntax || "cognition_verb"} negative_cognition_predicate`,
-    slots: ["negative_cognition_predicate", "negative_cognition_fragment"],
-    reason: "Phase 4 controlled grammar promotion: 知 is parser-active only inside the reviewed 唔知 negative-cognition fragment; 唔 remains a learner-visible func negator."
-  });
-  const children = [negator, promoted, ...particles];
-  return construction("NegativeCognitionFragment", "NegCognition", children, {
-    note: "Phase 4 controlled negative cognition fragment: 唔 + 知 as a stable short answer / learner fragment, with 唔 kept as func.",
-    slots: templateDerivedSlots("NegativeCognitionFragment", children),
-    trace: traceInfo("phase4_controlled_grammar_promotion", {
-      rule: "negator + cognition_predicate + particle?",
-      reason: "Promote only reviewed 唔知-style negative cognition fragments. The negator remains a learner-visible func token; this does not promote general cognition, speech, or stance domains.",
-      surfaces: children.map((node) => flattenSurface(node))
-    })
-  });
-}
+var createCognitionDetectors = require_fragments_content();
+var {
+  cognitionContentFrameFallback,
+  cognitionStatementFallback,
+  negativeCognitionFragmentFallback
+} = createCognitionDetectors({
+  aNotAQuestionFallback,
+  applyConstructionPatterns,
+  completionQuestionFallback,
+  construction,
+  desiderativeANotAQuestionFallback,
+  flattenSurface,
+  isToken,
+  nodeCanFillSlot,
+  nodeSurfaceMatches,
+  optionalSubjectOffset,
+  permissionANotAQuestionFallback,
+  phase4CognitionActiveTokenClone,
+  predicateOmissionProfileForHead,
+  templateConstructionFor,
+  templateDerivedSlots,
+  token,
+  traceInfo,
+  withoutTrailingParticles,
+  wrapCategorySubspans
+});
+var createOpinionDetectors = require_stance();
+var { opinionStanceFrameFallback } = createOpinionDetectors({
+  applyConstructionPatterns,
+  categorySubspanFor,
+  cleanSlots,
+  construction,
+  copulaClone,
+  firstToken,
+  flattenSurface,
+  isToken,
+  modalVPFromNodes,
+  nodeCanFillSlot,
+  nodeSlots,
+  nominalComplementFromNodes,
+  parserInactiveTokenClone,
+  phase4OpinionStanceActiveTokenClone,
+  predicateOmissionProfileForHead,
+  subjectStativePredicateClauseFallback,
+  templateConstructionFor,
+  templateDerivedSlots,
+  traceInfo,
+  withoutTrailingParticles,
+  wrapCategorySubspans
+});
+var createReportedSpeechDetectors = require_composition();
+var { reportedSpeechFrameFallback } = createReportedSpeechDetectors({
+  applyConstructionPatterns,
+  construction,
+  firstToken,
+  flattenSurface,
+  nodeCanFillSlot,
+  optionalSubjectOffset,
+  parserInactiveTokenClone,
+  phase4ReportedSpeechActiveTokenClone,
+  subjectStativePredicateClauseFallback,
+  templateConstructionFor,
+  templateDerivedSlots,
+  traceInfo,
+  withoutTrailingParticles,
+  wrapCategorySubspans
+});
 function transparentTopicContentFromNodes(nodes) {
   const compact = withoutIgnorableSpaceText(nodes || []);
   if (!compact.length) return null;
@@ -12133,532 +12820,6 @@ function incompletePotentialResultCandidate(core) {
       speech_event_use: "not_applicable",
       semantic_review_flags: ["needs_context_parse", "acceptability_potential_boundary"],
       not_claims: ["not_clean_acceptability_response", "not_fabricated_result_complement", "not_sentence_specific_surface_rule"]
-    })
-  });
-}
-function cognitionStatementFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  if (bareCore.length !== 2) return null;
-  const subject = bareCore[0];
-  const predicate = bareCore[1];
-  if (!nodeCanFillSlot(subject, "subject")) return null;
-  if (!nodeCanFillSlot(predicate, "cognition_predicate")) return null;
-  if (nodeCanFillSlot(predicate, "negative_cognition_predicate")) return null;
-  if (nodeCanFillSlot(predicate, "stance_predicate")) return null;
-  const promoted = phase4CognitionActiveTokenClone(predicate, {
-    label: predicate.label || "doing",
-    syntax: `${predicate.syntax || "cognition_verb"} cognition_statement_predicate`,
-    slots: ["cognition_statement_clause"],
-    reason: "Phase 4 controlled grammar promotion: bare subject + 知 is parser-active as a reviewed cognition statement, not a broad cognition-content frame."
-  });
-  const children = [subject, promoted, ...particles];
-  return construction("CognitionStatementClause", "Know", children, {
-    note: "Phase 4 controlled cognition statement: subject + 知 + optional particle.",
-    slots: templateDerivedSlots("CognitionStatementClause", children),
-    trace: traceInfo("phase4_controlled_grammar_promotion", {
-      rule: "subject + non_stance_cognition_predicate + particle?",
-      reason: "Promote only reviewed bare cognition statements such as 我知 / 我知呀. Content frames still require visible clause-like content, and desiderative 我想知 remains separate.",
-      predicate_omission_profile: "factive_cognition",
-      omission_status: "conventional_bare_statement",
-      complement_type: "proposition_or_interrogative_clause",
-      conventionality_status: "conventional_bare_statement",
-      polarity: "positive",
-      speech_event_use: "not_applicable",
-      not_claims: ["not_fabricated_content_clause", "not_forced_fragment_answer"],
-      surfaces: children.map((node) => flattenSurface(node))
-    })
-  });
-}
-function cognitionContentFrameFallback(core) {
-  let cursor = optionalSubjectOffset(core);
-  const subjectNodes = core.slice(0, cursor);
-  const negator = isToken(core[cursor], "唔") ? core[cursor++] : null;
-  if (core.length - cursor < 2) return null;
-  const cognition = core[cursor];
-  if (!nodeCanFillSlot(cognition, "cognition_predicate")) return null;
-  if (nodeCanFillSlot(cognition, "negative_cognition_predicate")) return null;
-  if (nodeCanFillSlot(cognition, "stance_predicate")) return null;
-  if (nodeCanFillSlot(cognition, "vp_complement_predicate") && !nodeSurfaceMatches(cognition, ["知"])) return null;
-  const rawContent = core.slice(cursor + 1);
-  const { core: contentCore, particles } = withoutTrailingParticles(rawContent);
-  const directClauseRelationEdge = contentCore.length === 1 && contentCore[0].kind === "construction" && contentCore[0].type === "ClauseRelationEdge" ? contentCore[0] : null;
-  if (!directClauseRelationEdge && contentCore.length < 2) return null;
-  const rawEmbeddedQuestion = directClauseRelationEdge ? null : aNotAQuestionFallback(contentCore) || permissionANotAQuestionFallback(contentCore) || desiderativeANotAQuestionFallback(contentCore) || completionQuestionFallback(contentCore);
-  const parsedContent = directClauseRelationEdge ? [directClauseRelationEdge] : rawEmbeddedQuestion ? [rawEmbeddedQuestion] : applyConstructionPatterns(contentCore);
-  const wrappedContent = parsedContent.length === 1 ? parsedContent : wrapCategorySubspans(contentCore);
-  let contentNode = directClauseRelationEdge;
-  if (!contentNode && (wrappedContent.length === 1 && wrappedContent[0].kind === "construction" && (nodeCanFillSlot(wrappedContent[0], "reported_content") || nodeCanFillSlot(wrappedContent[0], "clause") || nodeCanFillSlot(wrappedContent[0], "vp") || nodeCanFillSlot(wrappedContent[0], "predicate") || nodeCanFillSlot(wrappedContent[0], "question_fragment")))) {
-    contentNode = wrappedContent[0];
-  } else if (!contentNode) {
-    const generatedContent = templateConstructionFor(wrappedContent, [
-      "ANotAQuestion",
-      "OpinionStanceFrame",
-      "ReportedSpeech",
-      "ExistentialClause",
-      "NegatedExistentialClause",
-      "ExistentialQuestion",
-      "DesiderativeVP",
-      "ModalVP",
-      "CompletionQuestion"
-    ]);
-    if (generatedContent && (nodeCanFillSlot(generatedContent, "reported_content") || nodeCanFillSlot(generatedContent, "clause") || nodeCanFillSlot(generatedContent, "vp") || nodeCanFillSlot(generatedContent, "predicate") || nodeCanFillSlot(generatedContent, "question_fragment"))) contentNode = generatedContent;
-  }
-  if (!contentNode) return null;
-  const promotedCognition = phase4CognitionActiveTokenClone(cognition, {
-    label: cognition.label || "doing",
-    syntax: `${cognition.syntax || "cognition_verb"} non_stance_cognition_predicate`,
-    slots: ["non_stance_cognition_predicate"],
-    reason: "A1 broad cognition architecture: a 知-type predicate is parser-active when it licenses a visible declarative or interrogative content complement."
-  });
-  const children = [...subjectNodes, ...negator ? [negator] : [], promotedCognition, contentNode, ...particles];
-  const profile = predicateOmissionProfileForHead(flattenSurface(cognition));
-  return construction("CognitionContentFrame", "Know+Content", children, {
-    note: "Cognition content frame with visible polarity and a transparent declarative or interrogative complement.",
-    slots: templateDerivedSlots("CognitionContentFrame", children),
-    trace: traceInfo("phase4_controlled_grammar_promotion", {
-      rule: "subject? + polarity? + non_stance_cognition_predicate + content_clause + particle?",
-      reason: "The content remains an overt child construction. Negative polarity composes outside the cognition head, and embedded A-not-A content is not flattened into multiple roots.",
-      content_construction: contentNode.type,
-      predicate_omission_profile: profile ? profile.id : "factive_cognition",
-      omission_status: "overt_content_saturated",
-      complement_type: contentNode.type && String(contentNode.type).includes("Question") ? "interrogative_clause" : "proposition",
-      conventionality_status: "overt_content",
-      polarity: negator ? "negative" : "positive",
-      speech_event_use: "not_applicable",
-      surfaces: children.map((node) => flattenSurface(node)),
-      not_claims: ["not_hidden_content_clause", "not_stance_frame", "not_reported_speech"]
-    })
-  });
-}
-function splitOpinionTrailingParticles(core) {
-  const split = withoutTrailingParticles(core);
-  const bareCore = split.core.slice();
-  const particles = split.particles.slice();
-  if (particles.length && isToken(particles[0], "嘅") && bareCore.some((node) => isToken(node, "係"))) {
-    bareCore.push(particles.shift());
-  }
-  return { core: bareCore, particles };
-}
-function isOpinionFrameFocusModifier(node) {
-  return Boolean(node) && node.kind !== "text" && nodeCanFillSlot(node, "focus_adverb");
-}
-function splitLeadingOpinionContentModifiers(nodes) {
-  const modifiers = [];
-  let index = 0;
-  while (index < nodes.length && isOpinionFrameFocusModifier(nodes[index])) {
-    modifiers.push(nodes[index]);
-    index += 1;
-  }
-  return { modifiers, core: nodes.slice(index) };
-}
-function opinionComplementTokenClone(node, subtype) {
-  if (!node || node.kind !== "token") return node;
-  const surface = flattenSurface(node);
-  if (["係", "唔係"].includes(surface)) {
-    return parserInactiveTokenClone(node, {
-      label: "func",
-      pos: "function",
-      syntax: `${surface === "唔係" ? "negated_copula" : "copula"} complement_ellipsis_predicate`,
-      slots: cleanSlots([
-        surface === "唔係" ? "negated_copula" : "copula",
-        surface === "唔係" ? "negator" : "",
-        "predicate",
-        "reported_content",
-        "content_clause"
-      ]),
-      reason: `The visible ${surface} predicate is retained inside a ${subtype} complement-ellipsis fragment; no missing complement words are fabricated.`
-    });
-  }
-  if (["有", "冇"].includes(surface)) {
-    return parserInactiveTokenClone(node, {
-      label: "func",
-      pos: "function",
-      syntax: `${surface === "冇" ? "negated_existential" : "existential"} complement_ellipsis_predicate`,
-      slots: cleanSlots([
-        surface === "冇" ? "negated_existential" : "existential",
-        "predicate",
-        "reported_content",
-        "content_clause"
-      ]),
-      reason: `The visible ${surface} predicate is retained inside an existential-domain complement-ellipsis fragment; no missing domain is fabricated.`
-    });
-  }
-  return node;
-}
-function opinionComplementEllipsisFragment(nodes, subtype, missingSlot) {
-  const children = nodes.map((node) => opinionComplementTokenClone(node, subtype));
-  return construction("ComplementEllipsisFragment", "ComplementEllipsis", children, {
-    slots: cleanSlots([
-      "complement_ellipsis_fragment",
-      "fragment_answer",
-      "reported_content",
-      "content_clause",
-      "predicate",
-      "clause",
-      ...children.flatMap((node) => nodeSlots(node))
-    ]),
-    note: "Visible opinion content with a typed complement/domain omitted and recoverable only from discourse.",
-    trace: traceInfo("generative_template", {
-      construction_type: "ComplementEllipsisFragment",
-      template_family: "generative_template",
-      template: subtype === "existential_domain_ellipsis" ? ["existential_predicate!"] : nodes.length > 1 ? ["subject?", "copular_predicate!"] : ["copular_predicate!"],
-      assigned_slots: nodes.map((node, index) => {
-        if (index === 0 && nodes.length > 1 && nodeCanFillSlot(node, "subject")) return "subject";
-        return subtype === "existential_domain_ellipsis" ? "existential_predicate" : "copular_predicate";
-      }),
-      fragment_subtype: subtype,
-      context_requirement_status: "context_required",
-      missing_argument_slots: [missingSlot],
-      missing_slot_details: [{ slot: missingSlot, license_status: "unresolved" }],
-      antecedent_status: "not_observed",
-      discourse_license_not_observed: true,
-      surfaces: children.map((node) => flattenSurface(node)),
-      reason: "The overt predicate is structurally valid, but its selected complement/domain is absent and must remain linked to discourse rather than being invented.",
-      not_claims: ["not_context_free_complete_clause", "not_fabricated_complement", "not_lexicalized_whole_sentence"]
-    })
-  });
-}
-function nominalizedStativeComplementForOpinion(nodes) {
-  if (!nodes || nodes.length !== 2) return null;
-  const stative = nodes[0];
-  const linker = nodes[1];
-  if (!nodeCanFillSlot(stative, "stative_predicate") || !isToken(linker, "嘅")) return null;
-  const nominalizer = parserInactiveTokenClone(linker, {
-    label: "func",
-    pos: "function",
-    syntax: "nominalizer nominal_linker",
-    slots: ["nominal_linker"],
-    reason: "嘅 nominalizes the visible stative predicate inside a copular opinion complement; it is not treated as a sentence-final particle."
-  });
-  const children = [stative, nominalizer];
-  return construction("StativeNominalComplement", "AdjNP", children, {
-    slots: cleanSlots([
-      "stative_nominal_complement",
-      "copular_complement",
-      "np",
-      "predicate",
-      "stative_predicate",
-      "nominal_linker",
-      ...children.flatMap((node) => nodeSlots(node))
-    ]),
-    note: "Headless nominalized stative complement such as 真嘅 in 係真嘅.",
-    trace: traceInfo("generative_template", {
-      construction_type: "StativeNominalComplement",
-      template_family: "generative_template",
-      template: ["stative_predicate!", "nominal_linker!"],
-      assigned_slots: ["stative_predicate", "nominal_linker"],
-      complement_subtype: "headless_nominalized_stative",
-      surfaces: children.map((node) => flattenSurface(node)),
-      reason: "A visible stative predicate plus nominalizer forms the copular complement without inventing a silent noun head.",
-      not_claims: ["not_overt_head_noun", "not_sentence_final_particle_use"]
-    })
-  });
-}
-function opinionCopularContentNode(nodes) {
-  if (!nodes || nodes.length < 2 || !isToken(nodes[0], "係")) return null;
-  const complementNodes = nodes.slice(1);
-  const complement = nominalizedStativeComplementForOpinion(complementNodes) || nominalComplementFromNodes(complementNodes) || categorySubspanFor(complementNodes, ["DegreeStativePredicate", "NegatedStativePredicate", "StativePredicate"]);
-  if (!complement) return null;
-  const copula = copulaClone(
-    nodes[0],
-    "copula subordinate_opinion_copula",
-    ["copula"],
-    "係 heads a visible copular proposition inside the opinion complement. Its understood proposition subject is licensed by the embedding stance frame, not fabricated as a token."
-  );
-  const children = [copula, complement];
-  return construction("CopularRelationFrame", "Relation", children, {
-    slots: cleanSlots([
-      "copular_relation_frame",
-      "copular_clause",
-      "copula",
-      "copular_complement",
-      "reported_content",
-      "content_clause",
-      "predicate",
-      "clause",
-      ...children.flatMap((node) => nodeSlots(node))
-    ]),
-    note: "Subordinate copular proposition used as visible opinion content.",
-    trace: traceInfo("generative_template", {
-      construction_type: "CopularRelationFrame",
-      template_family: "generative_template",
-      template: ["copula!", "copular_complement!"],
-      assigned_slots: ["copula", "copular_complement"],
-      subject_status: "null_licensed_by_opinion_content",
-      context_requirement_status: "context_not_required",
-      missing_argument_slots: [],
-      antecedent_status: "not_applicable",
-      surfaces: children.map((node) => flattenSurface(node)),
-      reason: "Inside a stance complement, 係 plus a visible nominalized/stative complement forms a complete proposition without requiring an invented overt subject.",
-      not_claims: ["not_standalone_context_free_subject_claim", "not_complement_ellipsis"]
-    })
-  });
-}
-function contentNodeForOpinionStanceFrame(contentCore) {
-  if (!contentCore.length) return null;
-  const split = splitLeadingOpinionContentModifiers(contentCore);
-  const leadingModifiers = split.modifiers;
-  const core = split.core;
-  if (!core.length) return null;
-  if (core.length === 1 && core[0].kind === "construction" && core[0].type === "ClauseRelationEdge") {
-    return { leadingModifiers, contentNode: core[0] };
-  }
-  if (core.length === 1 && ["有", "冇"].includes(flattenSurface(core[0]))) {
-    return {
-      leadingModifiers,
-      contentNode: opinionComplementEllipsisFragment(core, "existential_domain_ellipsis", "existential_domain")
-    };
-  }
-  const finalSurface = flattenSurface(core[core.length - 1]);
-  const copularEllipsis = ["係", "唔係"].includes(finalSurface) && (core.length === 1 || core.length === 2 && nodeCanFillSlot(core[0], "subject"));
-  if (copularEllipsis) {
-    return {
-      leadingModifiers,
-      contentNode: opinionComplementEllipsisFragment(core, "copular_complement_ellipsis", "copular_complement")
-    };
-  }
-  const copularContent = opinionCopularContentNode(core);
-  if (copularContent) return { leadingModifiers, contentNode: copularContent };
-  const modalContent = modalVPFromNodes(core);
-  if (modalContent) return { leadingModifiers, contentNode: modalContent };
-  const stativeContent = categorySubspanFor(core, [
-    "DegreeStativePredicate",
-    "NegatedStativePredicate",
-    "DegreeModifiedLexicalStative",
-    "StativePredicate"
-  ]);
-  if (stativeContent) return { leadingModifiers, contentNode: stativeContent };
-  const fullyWrappedContent = applyConstructionPatterns(core);
-  if (fullyWrappedContent.length === 1 && fullyWrappedContent[0].kind === "construction" && [
-    "ClauseSpan",
-    "TopicComment",
-    "CopularRelationFrame",
-    "CopularIdentificationFrame",
-    "ExistentialClause",
-    "NegatedExistentialClause",
-    "DesiderativeVP",
-    "ModalVP",
-    "CompletionQuestion",
-    "ANotAQuestion",
-    "PotentialResultVP",
-    "PerfectiveVP",
-    "PostverbalZoPerfectiveVP",
-    "DirectionalMotionVP",
-    "DegreeStativePredicate",
-    "NegatedStativePredicate",
-    "StativePredicate",
-    "ComplementEllipsisFragment",
-    "ClauseRelationEdge"
-  ].includes(fullyWrappedContent[0].type)) {
-    return { leadingModifiers, contentNode: fullyWrappedContent[0] };
-  }
-  const wrappedContent = wrapCategorySubspans(core);
-  const subjectStative = subjectStativePredicateClauseFallback(wrappedContent);
-  if (subjectStative) return { leadingModifiers, contentNode: subjectStative };
-  if (wrappedContent.length === 1 && wrappedContent[0].kind === "construction" && [
-    "ClauseSpan",
-    "TopicComment",
-    "CopularRelationFrame",
-    "CopularIdentificationFrame",
-    "ExistentialClause",
-    "NegatedExistentialClause",
-    "DesiderativeVP",
-    "ModalVP",
-    "CompletionQuestion",
-    "DegreeStativePredicate",
-    "NegatedStativePredicate",
-    "StativePredicate",
-    "ComplementEllipsisFragment",
-    "ClauseRelationEdge"
-  ].includes(wrappedContent[0].type)) {
-    return { leadingModifiers, contentNode: wrappedContent[0] };
-  }
-  const generatedContent = templateConstructionFor(wrappedContent, [
-    "ExistentialClause",
-    "NegatedExistentialClause",
-    "DesiderativeVP",
-    "ModalVP",
-    "CompletionQuestion"
-  ]);
-  if (generatedContent) return { leadingModifiers, contentNode: generatedContent };
-  return null;
-}
-function opinionStanceFrameFallback(core) {
-  const split = splitOpinionTrailingParticles(core);
-  const bareCore = split.core;
-  const particles = split.particles;
-  if (bareCore.length < 3) return null;
-  let cursor = 0;
-  const subject = nodeCanFillSlot(bareCore[cursor], "subject") ? bareCore[cursor++] : null;
-  const preStanceModifiers = [];
-  while (cursor < bareCore.length && isOpinionFrameFocusModifier(bareCore[cursor])) {
-    preStanceModifiers.push(bareCore[cursor]);
-    cursor += 1;
-  }
-  const stancePredicate = bareCore[cursor];
-  const stanceToken = firstToken(stancePredicate);
-  if (!stanceToken || !nodeCanFillSlot(stancePredicate, "stance_predicate")) return null;
-  const profile = predicateOmissionProfileForHead(stanceToken.surface);
-  cursor += 1;
-  if (cursor >= bareCore.length) return null;
-  const rawStanceContent = bareCore.slice(cursor);
-  const contentAnalysis = rawStanceContent.length === 1 && rawStanceContent[0].kind === "construction" && rawStanceContent[0].type === "ClauseRelationEdge" ? { leadingModifiers: [], contentNode: rawStanceContent[0] } : contentNodeForOpinionStanceFrame(rawStanceContent);
-  if (!contentAnalysis || !contentAnalysis.contentNode) return null;
-  const contentNode = contentAnalysis.contentNode;
-  const postStanceModifiers = contentAnalysis.leadingModifiers || [];
-  const promotedStance = phase4OpinionStanceActiveTokenClone(stancePredicate, {
-    label: stancePredicate.label || "doing",
-    syntax: `${stanceToken.syntax || "cognition_opinion_predicate"} opinion_stance_predicate`,
-    slots: ["opinion_stance_frame"],
-    reason: "A1 broad stance architecture: a registered stance or belief predicate is parser-active when it licenses visible reviewed content."
-  });
-  const children = [
-    ...subject ? [subject] : [],
-    ...preStanceModifiers,
-    promotedStance,
-    ...postStanceModifiers,
-    contentNode,
-    ...particles
-  ];
-  const assignedSlots = [
-    ...subject ? ["subject"] : [],
-    ...preStanceModifiers.map(() => "focus_adverb"),
-    "stance_predicate",
-    ...postStanceModifiers.map(() => "focus_adverb"),
-    "reported_content",
-    ...particles.map(() => "particle")
-  ];
-  const contentTrace = contentNode.trace || {};
-  const contextRequirementStatus = contentTrace.context_requirement_status || "context_not_required";
-  const missingArgumentSlots = Array.isArray(contentTrace.missing_argument_slots) ? contentTrace.missing_argument_slots.slice() : [];
-  const missingSlotDetails = Array.isArray(contentTrace.missing_slot_details) ? contentTrace.missing_slot_details.map((item) => ({ ...item })) : missingArgumentSlots.map((slot) => ({ slot, license_status: "unresolved" }));
-  return construction("OpinionStanceFrame", "Opinion/Stance", children, {
-    note: "Opinion, evaluation, or belief frame with a visible transparent proposition or typed complement ellipsis.",
-    slots: cleanSlots([
-      ...templateDerivedSlots("OpinionStanceFrame", children),
-      "opinion_stance_frame",
-      "stance_predicate",
-      "reported_content",
-      "content_clause",
-      "predicate",
-      "clause"
-    ]),
-    trace: traceInfo("generative_template", {
-      construction_type: "OpinionStanceFrame",
-      template_family: "generative_template",
-      template: ["subject?", "focus_adverb?", "stance_predicate!", "focus_adverb?", "reported_content!", "particle?"],
-      constraints: {
-        visible_reviewed_content: true,
-        stance_predicate_surfaces: [stanceToken.surface],
-        focus_modifier_slot: "focus_adverb"
-      },
-      assigned_slots: assignedSlots,
-      content_construction: contentNode.type,
-      pre_stance_modifier_surfaces: preStanceModifiers.map((node) => flattenSurface(node)),
-      post_stance_modifier_surfaces: postStanceModifiers.map((node) => flattenSurface(node)),
-      context_requirement_status: contextRequirementStatus,
-      missing_argument_slots: missingArgumentSlots,
-      missing_slot_details: missingSlotDetails,
-      antecedent_status: contextRequirementStatus === "context_required" ? "not_observed" : "not_applicable",
-      discourse_license_not_observed: contextRequirementStatus === "context_required",
-      attached_particles: particles.map((node) => flattenSurface(node)),
-      surfaces: children.map((node) => flattenSurface(node)),
-      predicate_omission_profile: profile ? profile.id : "opinion_evaluation",
-      reason: "The reusable stance frame is keyed by registered stance-predicate affordance, not one hard-coded surface. It wraps—but does not replace—the visible content construction.",
-      not_claims: ["not_reported_speech", "not_action_stative_vp_overcapture", "not_fabricated_opinion_content"]
-    })
-  });
-}
-function contentNodeForReportedSpeech(contentCore) {
-  if (!contentCore.length) return null;
-  if (contentCore.length === 1 && contentCore[0].kind === "construction" && contentCore[0].type === "ClauseRelationEdge") {
-    return contentCore[0];
-  }
-  if (contentCore.length === 1 && contentCore[0].kind === "token") {
-    const contentToken = contentCore[0];
-    if (["有", "冇"].includes(flattenSurface(contentToken)) && (nodeCanFillSlot(contentToken, "existential") || nodeCanFillSlot(contentToken, "negated_existential"))) {
-      return parserInactiveTokenClone(contentToken, {
-        label: contentToken.label || contentToken.role || "func",
-        syntax: `${contentToken.syntax || "predicate"} reported_content_predicate`,
-        slots: [...contentToken.slots || [], "reported_content", "predicate"],
-        reason: "A visible short predicate is functioning as the reported content of 話; no standalone response-fragment reading or hidden clause is introduced."
-      });
-    }
-  }
-  const parsedContent = applyConstructionPatterns(contentCore);
-  const wrappedContent = parsedContent.length === 1 ? parsedContent : wrapCategorySubspans(contentCore);
-  const subjectStative = subjectStativePredicateClauseFallback(wrappedContent);
-  if (subjectStative) return subjectStative;
-  if (wrappedContent.length === 1 && wrappedContent[0].kind === "construction" && (nodeCanFillSlot(wrappedContent[0], "reported_content") || nodeCanFillSlot(wrappedContent[0], "clause") || nodeCanFillSlot(wrappedContent[0], "vp") || nodeCanFillSlot(wrappedContent[0], "predicate"))) {
-    return wrappedContent[0];
-  }
-  if (wrappedContent.length === 1 && wrappedContent[0].kind === "token") {
-    const contentToken = wrappedContent[0];
-    if (nodeCanFillSlot(contentToken, "predicate") || nodeCanFillSlot(contentToken, "existential") || nodeCanFillSlot(contentToken, "negated_existential") || nodeCanFillSlot(contentToken, "vp") || nodeCanFillSlot(contentToken, "clause")) {
-      return parserInactiveTokenClone(contentToken, {
-        label: contentToken.label || contentToken.role || "func",
-        syntax: `${contentToken.syntax || "predicate"} reported_content_predicate`,
-        slots: [...contentToken.slots || [], "reported_content", "predicate"],
-        reason: "A visible short predicate is functioning as the reported content of 話; no hidden clause or lexicalized whole-sentence formula is introduced."
-      });
-    }
-  }
-  const generatedContent = templateConstructionFor(wrappedContent, [
-    "CognitionContentFrame",
-    "OpinionStanceFrame",
-    "ExistentialClause",
-    "NegatedExistentialClause",
-    "DesiderativeVP",
-    "ModalVP",
-    "CompletionQuestion"
-  ]);
-  if (generatedContent && (nodeCanFillSlot(generatedContent, "reported_content") || nodeCanFillSlot(generatedContent, "clause") || nodeCanFillSlot(generatedContent, "vp") || nodeCanFillSlot(generatedContent, "predicate"))) return generatedContent;
-  return null;
-}
-function reportedSpeechFrameFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  const offset = optionalSubjectOffset(bareCore);
-  if (bareCore.length - offset < 2) return null;
-  const speechVerb = bareCore[offset];
-  const speechToken = firstToken(speechVerb);
-  if (!speechToken || speechToken.surface !== "話") return null;
-  if (!nodeCanFillSlot(speechVerb, "speech_verb")) return null;
-  const contentNode = contentNodeForReportedSpeech(bareCore.slice(offset + 1));
-  if (!contentNode) return null;
-  const promotedSpeech = phase4ReportedSpeechActiveTokenClone(speechVerb, {
-    label: speechVerb.label || "doing",
-    syntax: `${speechToken.syntax || "speech_reporting_verb"} reported_speech_predicate`,
-    slots: ["reported_speech"],
-    reason: "Phase 4 controlled grammar promotion: 話 is parser-active only when it introduces a visible reviewed reported-content frame."
-  });
-  const children = [
-    ...bareCore.slice(0, offset),
-    promotedSpeech,
-    contentNode,
-    ...particles
-  ];
-  const assignedSlots = [
-    ...bareCore.slice(0, offset).map(() => "subject"),
-    "speech_verb",
-    "reported_content",
-    ...particles.map(() => "particle")
-  ];
-  return construction("ReportedSpeech", "Reported", children, {
-    note: "Reported speech frame: optional subject + speech-reporting verb + visible reviewed reported content.",
-    slots: templateDerivedSlots("ReportedSpeech", children),
-    trace: traceInfo("generative_template", {
-      construction_type: "ReportedSpeech",
-      template_family: "generative_template",
-      template: ["subject?", "speech_verb!", "reported_content!", "particle?"],
-      constraints: {
-        visible_reviewed_content: true,
-        speech_verb_surfaces: ["話"]
-      },
-      assigned_slots: assignedSlots,
-      content_construction: contentNode.type,
-      attached_particles: particles.map((node) => flattenSurface(node)),
-      surfaces: children.map((node) => flattenSurface(node)),
-      reason: "The frame is now reported as a named reusable generative template; the controlled parser-active speech predicate remains scoped to this approved construction."
     })
   });
 }
